+++ CHANGELOG.md
@@ -67,7 +67,8 @@
 - Remove deprecated remove() from Pim/Bundle/CatalogBundle/Manager/CategoryManager
 - Remove deprecated remove() from Pim/Bundle/CatalogBundle/Manager/FamilyManager
 - Remove deprecated remove() from Pim/Bundle/CatalogBundle/Manager/GroupManager
-- Change arguments of Pim\Bundle\EnrichBundle\Controller\AssociationController to use AssociationTypeRepositoryInterface, ProductRepositoryInterface, ProductBuilderInterface, EngineInterface
+- Change arguments of Pim/Bundle/EnrichBundle/Controller/AssociationController to use AssociationTypeRepositoryInterface, ProductRepositoryInterface, ProductBuilderInterface, EngineInterface
+- Remove arguments ChannelRepositoryInterface, LocaleRepositoryInterface, add argument AttributeValuesResolver in Pim/Bundle/CatalogBundle/Builder/ProductBuilder constructor
 
 # 1.3.x